+++ CHANGES.rst
@@ -9,6 +9,7 @@
 - Make pypeit_show_2dspec (somewhat) backwards compatible.
 - Added the option to disable strict version checking for 1d coadds.
 - Sensitivity function masking and output updates
+- Updates to reduction parameters for LDT/DeVeny
 
 1.12.2 (29 Mar 2023)
 --------------------